+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+2.1.2 (2017-10-06)
+------------------
+
+    * Fix briefy.leads quote events' handlers (ericof).
+    * Add Additional Info and Category to Quote actions (ericof).
+
 2.1.1 (2017-09-05)
 ------------------